moving grid constants from global_constants.py to grid.py
