+++ fv3core/utils/global_constants.py
@@ -35,11 +35,3 @@ ql0_max = 2.0e-3  # max cloud water value (auto converted to rain)
 t_sub = 184.0  # min temp for sublimation of cloud ice
 DC_ICE = C_LIQ - C_ICE
 LI0 = HLF - DC_ICE * TICE
-
-# grid condtants
-# TODO: move these into the fv3core.grid namespace
-LON_OR_LAT_DIM = "lon_or_lat"
-TILE_DIM = "tile"
-CARTESIAN_DIM = "xyz_direction"
-N_TILES = 6
-RIGHT_HAND_GRID = False
